fix(epic16): review follow-ups on the Story 16.2 backend wiring (PR #451)
Silent int narrowing on the bulk snapshot finders (SScholefield). Bridge,
culvert and silviculture-location ids are `long` on their entity records, and
each service cast them to `int` on the way into the shared finder — a silent
truncation past Integer.MAX_VALUE that would have surfaced as *empty original
values*, not as an error. The shared boundary now speaks Long: Row.parentId
and every `IN`-list parameter are `Long`, the casts are gone, and the callers
whose own ids are `int` widen losslessly instead. Row.detailId stays Integer —
its callers are Integer end to end with no cast, so widening it would only
manufacture Integer/Long key mismatches.

Non-deterministic general comment on Schedule 6 (SScholefield). Confirmed:
findRoadRecords orders by ROAD_MAINTENANCE_REPORT_ID and findRoadRecordSnapshots
did not, so the "last row wins" rule that legacy applies to the replicated
general comment was reading whichever row the plan happened to return last.
Both queries now order by the same column, so the submitted comment comes from
the same row of the family as the current one.

CampOriginals.build() -> builder() (SScholefield) — it returns a builder, and
`originals.build()...build()` read as the same call twice.

Duplicate ResultSet null helpers (SScholefield) -> ca.bc.gov.nrs.ilcr.util
.ResultSetUtil. Three copies of the getInt/wasNull pair, not two: Schedule 8
and Schedule 10 as reported, plus the shared cost-detail snapshot mapper.
Schedule 1 and Schedule 3 keep theirs — those read the column as a BigDecimal
and narrow it, a different strategy rather than a fourth copy.

Test coverage for what the refactor touched. Schedules 6, 7A, 9 and 10 had NO
unit test asserting a cost snapshot reaches the served row, which is exactly
the join whose key type changed here — a mis-keyed lookup would have silently
emptied the indicators. Thirteen tests now pin, per schedule: Draft (no map,
no query issued), the report-snapshot/cost-snapshot join on the record id with
values AND tooltips, and the empty-but-not-null case. This is the same gap
paulushcgcj flagged for Schedules 2-4 on #450, closed for the other half.

+++ backend/src/test/java/ca/bc/gov/nrs/ilcr/schedule1/Schedule1OriginalValuesTest.java
@@ -91,9 +91,21 @@ class Schedule1OriginalValuesTest {
         .thenReturn(
             List.of(
                 new CostDetailSnapshotRepository.Row(
-                    7101, SUMMARY_ID, CODE_STANDING_TREE, new BigDecimal("60000"), 600, null, null),
+                    7101,
+                    (long) SUMMARY_ID,
+                    CODE_STANDING_TREE,
+                    new BigDecimal("60000"),
+                    600,
+                    null,
+                    null),
                 new CostDetailSnapshotRepository.Row(
-                    7102, SUMMARY_ID, CODE_SUBTOTAL, new BigDecimal("60000"), 98_000, null, null)));
+                    7102,
+                    (long) SUMMARY_ID,
+                    CODE_SUBTOTAL,
+                    new BigDecimal("60000"),
+                    98_000,
+                    null,
+                    null)));
     when(summarySnapshots.findBySummaryId(SUMMARY_ID))
         .thenReturn(
             Optional.of(
@@ -199,7 +211,7 @@ class Schedule1OriginalValuesTest {
         .thenReturn(
             List.of(
                 new CostDetailSnapshotRepository.Row(
-                    7101, SUMMARY_ID, CODE_STANDING_TREE, null, null, null, null)));
+                    7101, (long) SUMMARY_ID, CODE_STANDING_TREE, null, null, null, null)));
     when(summarySnapshots.findBySummaryId(SUMMARY_ID))
         .thenReturn(Optional.of(new ReportSummarySnapshotRepository.Snapshot(null, null, null)));
 
@@ -220,9 +232,9 @@ class Schedule1OriginalValuesTest {
         .thenReturn(
             List.of(
                 new CostDetailSnapshotRepository.Row(
-                    7201, SUMMARY_ID, 19, new BigDecimal("12345"), null, null, null),
+                    7201, (long) SUMMARY_ID, 19, new BigDecimal("12345"), null, null, null),
                 new CostDetailSnapshotRepository.Row(
-                    7202, SUMMARY_ID, 19, null, 1_200, "Aerial survey", null)));
+                    7202, (long) SUMMARY_ID, 19, null, 1_200, "Aerial survey", null)));
     when(summarySnapshots.findBySummaryId(SUMMARY_ID)).thenReturn(Optional.empty());
 
     Schedule1Response doc = served("S");
